Remove extra directories in installation dir for common
When using git to install the common subproject, and not using meson wrapper (mostly because I couldn't get it to work), the existing build file copies the entire repository into the  installation dir, including the tests and git folders. This change installs the headers only to (/usr/local)/include/common.