Add Dockerfile for containerised build
Added repeatable build recipe, using an alpine image, as the repository uses features only available in a recent version of Meson. As it builds to a static library, we just copy out the built files for the architecture into a plain image, and these can be installed to /usr/local from another image, or by extracting to a local filesystem.